+++ zerver/tests/test_events.py
@@ -206,7 +206,9 @@ from zerver.lib.event_schema import (
     check_realm_domains_add,
     check_realm_domains_change,
     check_realm_domains_remove,
+    check_realm_emoji_add,
     check_realm_emoji_update,
+    check_realm_emoji_update_one,
     check_realm_export,
     check_realm_export_consent,
     check_realm_linkifiers,
@@ -359,6 +361,7 @@ class BaseAction(ZulipTestCase):
         archived_channels: bool = False,
         allow_empty_topic_name: bool = True,
         simplified_presence_events: bool = False,
+        individual_emoji_changes: bool = True,
     ) -> Iterator[list[dict[str, Any]]]:
         """
         Make sure we have a clean slate of client descriptors for these tests.
@@ -392,6 +395,7 @@ class BaseAction(ZulipTestCase):
                 include_deactivated_groups=include_deactivated_groups,
                 archived_channels=archived_channels,
                 simplified_presence_events=simplified_presence_events,
+                individual_emoji_changes=individual_emoji_changes,
             )
         )
 
@@ -3342,10 +3346,26 @@ class NormalActionsTest(BaseAction):
             check_add_realm_emoji(
                 self.user_profile.realm, "my_emoji", author, img_file, "image/png"
             )
-
-        check_realm_emoji_update("events[0]", events[0])
+        check_realm_emoji_add("events[0]", events[0])
 
         with self.verify_action() as events:
+            do_remove_realm_emoji(
+                self.user_profile.realm, "my_emoji", acting_user=self.user_profile
+            )
+        check_realm_emoji_update_one("events[0]", events[0])
+
+    def test_realm_emoji_events_legacy(self) -> None:
+        author = self.example_user("iago")
+        with (
+            get_test_image_file("img.png") as img_file,
+            self.verify_action(individual_emoji_changes=False) as events,
+        ):
+            check_add_realm_emoji(
+                self.user_profile.realm, "my_emoji", author, img_file, "image/png"
+            )
+        check_realm_emoji_update("events[0]", events[0])
+
+        with self.verify_action(individual_emoji_changes=False) as events:
             do_remove_realm_emoji(
                 self.user_profile.realm, "my_emoji", acting_user=self.user_profile
             )